requirements.txt
asgiref==3.8.1
Django==5.0.8
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mssql-django==1.5
pillow==10.4.0
PyJWT==2.9.0
pyodbc==5.1.0
pytz==2024.1
sqlparse==0.5.1
typing_extensions==4.12.2
